requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
Django==5.0
django-crispy-forms==2.1
flake8==6.1.0
mccabe==0.7.0
Pillow==10.1.0
pycodestyle==2.11.1
pyflakes==3.1.0
sqlparse==0.4.4
typing_extensions==4.9.0
